♻️ Extract method (#1091)
